robot framework/requirements.txt
robotframework>=5
robotframework-faker==5.0.0
robotframework-seleniumlibrary==5.1.3
robotframework-debuglibrary==2.2.2
robotframework-requests==0.9.1
robotframework-pabot==2.0.0
robotframework-pdf2textlibrary==1.0.1
robotframework-jsonlibrary
pyyaml==5.4.1
xlrd==2.0.1
pandas==1.3.0
openpyxl==3.0.5